ciao-sdk: Add supporting functions in helpers.go
